pkg/chat_model/qwen.py

import os
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI

load_dotenv()

chat_model = ChatOpenAI(
    api_key=os.getenv("DASHSCOPE_API_KEY"),
    base_url="https://dashscope.aliyuncs.com/compatible-mode/v1",
    model="qwen3-max",
)